refactor(vscode): separate URI handling from commands
