+++ packages/open-collaboration-vscode/src/extension.ts
@@ -16,6 +16,7 @@ import { Commands } from './commands.js';
 import { Fetch } from './collaboration-connection-provider.js';
 import fetch from 'node-fetch';
 import { ChatWebview } from './chat-webview/chat-webview.js';
+import { JoinUriHandler } from './join-uri-handler.js';
 
 initializeProtocol({
     cryptoModule: crypto.webcrypto
@@ -26,6 +27,7 @@ export async function activate(context: vscode.ExtensionContext) {
     container.bind(Fetch).toConstantValue(fetch);
     const commands = container.get(Commands);
     commands.initialize();
+    container.get(JoinUriHandler).initialize();
     container.get(ChatWebview).register();
     const roomService = container.get(CollaborationRoomService);
